Updates from saarCTF 2023

+++ ci/buildscripts/post-install.sh
@@ -5,7 +5,7 @@ SCRIPTPATH="$(cd "$(dirname "$BASH_SOURCE")" && pwd)"
 source "${SCRIPTPATH}/../include/detect-docker"
 
 # Docker only - install systemd handler, remove the rest
-if detect-docker; then
+if detect-docker && [[ -z "${NO_DOCKER_SYSTEMD:-}" ]]; then
   cp /opt/gamelib/ci/docker-systemd/* /opt/
   rm -rf /opt/gamelib /opt/service
   rm -rf /tmp/*
